fix: separate out client components and add metadata
